updated hw2
