+++ .idea/compiler.xml
@@ -6,6 +6,8 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
+        <module name="hw20" />
+        <module name="hw21" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
