+++ .idea/misc.xml
@@ -10,6 +10,8 @@
         <option value="$PROJECT_DIR$/Projects/pom.xml" />
         <option value="$PROJECT_DIR$/Projects/untitled/pom.xml" />
         <option value="$PROJECT_DIR$/java-core/pom.xml" />
+        <option value="$PROJECT_DIR$/Coding/hw20/pom.xml" />
+        <option value="$PROJECT_DIR$/Coding/hw21/pom.xml" />
       </list>
     </option>
     <option name="ignoredFiles">
