--- Coding/Parking Lot/parkinglot/Ticket.java
@@ -1,27 +1,0 @@
-package parkinglot;
-
-public class Ticket {
-    private String ticketId;
-    private Vehicle vehicle;
-    private ParkingSpot spot;
-    private long entryTime;
-
-    public Ticket(Vehicle vehicle, ParkingSpot spot) {
-        this.ticketId = "TICKET-" + System.currentTimeMillis();
-        this.vehicle = vehicle;
-        this.spot = spot;
-        this.entryTime = System.currentTimeMillis();
-    }
-
-    public String getTicketId() {
-        return ticketId;
-    }
-
-    public Vehicle getVehicle() {
-        return vehicle;
-    }
-
-    public ParkingSpot getSpot() {
-        return spot;
-    }
-}
