--- Coding/Parking Lot/parkinglot/Vehicle.java
@@ -1,23 +1,0 @@
-package parkinglot;
-
-public class Vehicle {
-    private String licensePlate;
-    private VehicleType type;
-
-    public Vehicle(String licensePlate, VehicleType type) {
-        this.licensePlate = licensePlate;
-        this.type = type;
-    }
-
-    public String getLicensePlate() {
-        return licensePlate;
-    }
-
-    public VehicleType getType() {
-        return type;
-    }
-}
-
-enum VehicleType {
-    CAR, MOTORCYCLE
-}
