+++ Coding/hw20/src/main/java/Vehicle.java
@@ -1,0 +1,21 @@
+public class Vehicle {
+    private String licensePlate;
+    private VehicleType type;
+
+    public Vehicle(String licensePlate, VehicleType type) {
+        this.licensePlate = licensePlate;
+        this.type = type;
+    }
+
+    public String getLicensePlate() {
+        return licensePlate;
+    }
+
+    public VehicleType getType() {
+        return type;
+    }
+}
+
+enum VehicleType {
+    CAR, MOTORCYCLE
+}
