+++ Coding/hw21/src/main/java/notification/exception/NotificationException.java
@@ -1,0 +1,7 @@
+package notification.exception;
+
+public class NotificationException extends Exception{
+    public NotificationException(String message) {
+        super(message);
+    }
+}
