+++ Coding/hw21/src/main/java/notification/model/NotificationType.java
@@ -1,0 +1,5 @@
+package notification.model;
+
+public enum NotificationType {
+    EMAIL, SMS, WHATSAPP
+}
