+++ Coding/hw21/src/main/java/notification/sender/NotificationSender.java
@@ -1,0 +1,8 @@
+package notification.sender;
+
+import notification.exception.NotificationException;
+import notification.model.User;
+
+public interface NotificationSender {
+    void sendNotification(User user, String message) throws NotificationException;
+}
