--- Coding/parkinglot/parkinglot.iml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager">
-    <output url="file://$MODULE_DIR$/out/production/parkinglot" />
-    <output-test url="file:///test/parkinglot/Users/qzeng/Desktop/chuwa/Chuwa20241209/Coding/parkinglot/out/test/parkinglot" />
-    <exclude-output />
-    <content url="file://$MODULE_DIR$/..">
-      <sourceFolder url="file://$MODULE_DIR$/../Parking Lot" isTestSource="false" />
-    </content>
-    <orderEntry type="jdk" jdkName="openjdk-19" jdkType="JavaSDK" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-</module>
