--- java-core/src/main/java/com/chuwa/learn/Main.java
@@ -1,7 +1,0 @@
-package com.chuwa.learn;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
